refactor(selfupdate): define the updater launch contract
