+++ backend/internal/integration/updatecli/client_test.go
@@ -6,6 +6,8 @@ import (
 	"path/filepath"
 	"testing"
 	"time"
+
+	serviceselfupdate "github.com/futrx-com/remote.futrx.com/internal/service/selfupdate"
 )
 
 func TestStartUpdaterSelectsReleaseScript(t *testing.T) {
@@ -34,7 +36,14 @@ func TestStartUpdaterSelectsReleaseScript(t *testing.T) {
 			logPath := filepath.Join(installDir, "run.log")
 			donePath := filepath.Join(installDir, "done.json")
 			progressPath := filepath.Join(installDir, "progress.json")
-			if _, err := (Client{}).StartUpdater(installDir, "0.4.2", test.kind, logPath, donePath, progressPath); err != nil {
+			if _, err := (Client{}).StartUpdater(serviceselfupdate.UpdaterLaunch{
+				InstallDir:   installDir,
+				Target:       "0.4.2",
+				Kind:         serviceselfupdate.UpdateKind(test.kind),
+				LogPath:      logPath,
+				DonePath:     donePath,
+				ProgressPath: progressPath,
+			}); err != nil {
 				t.Fatal(err)
 			}
 			waitForDone(t, donePath)
@@ -52,7 +61,14 @@ func TestStartUpdaterSelectsReleaseScript(t *testing.T) {
 
 func TestStartUpdaterRejectsUnknownKind(t *testing.T) {
 	dir := t.TempDir()
-	if _, err := (Client{}).StartUpdater(dir, "0.4.2", "surprise", filepath.Join(dir, "log"), filepath.Join(dir, "done"), filepath.Join(dir, "progress")); err == nil {
+	if _, err := (Client{}).StartUpdater(serviceselfupdate.UpdaterLaunch{
+		InstallDir:   dir,
+		Target:       "0.4.2",
+		Kind:         "surprise",
+		LogPath:      filepath.Join(dir, "log"),
+		DonePath:     filepath.Join(dir, "done"),
+		ProgressPath: filepath.Join(dir, "progress"),
+	}); err == nil {
 		t.Fatal("StartUpdater accepted an unknown update kind")
 	}
 }
