+++ backend/internal/service/selfupdate/service.go
@@ -20,13 +20,6 @@ import (
 	"time"
 )
 
-// HostClient is implemented by integration/updatecli.
-type HostClient interface {
-	ListRemoteTags(ctx context.Context, installDir string) ([]string, error)
-	StartUpdater(installDir, tag, kind, logPath, donePath, progressPath string) (int, error)
-	ProcessAlive(pid int) bool
-}
-
 var (
 	ErrUpdateInProgress = errors.New("an update is already running")
 	ErrNoReleaseTag     = errors.New("no release tags found on origin")
@@ -37,16 +30,6 @@ const (
 	stateDirName = "self-update"
 	logTailBytes = 16 * 1024
 	runFileMode  = 0o600
-)
-
-// UpdateKind selects the smallest safe deployment path for a release.
-// Application updates stay within the installed major/minor release line;
-// crossing either boundary requires full infrastructure convergence.
-type UpdateKind string
-
-const (
-	UpdateKindApplication    UpdateKind = "application"
-	UpdateKindInfrastructure UpdateKind = "infrastructure"
 )
 
 type Service struct {
@@ -66,61 +49,6 @@ func New(currentVersion, installDir, dataDir string, host HostClient) *Service {
 		stateDir:       filepath.Join(dataDir, stateDirName),
 		host:           host,
 	}
-}
-
-// CheckResult is the outcome of one tag lookup against origin.
-type CheckResult struct {
-	CheckedAt       int64      `json:"checkedAt"`
-	LatestTag       string     `json:"latestTag,omitempty"`
-	UpdateAvailable bool       `json:"updateAvailable"`
-	UpdateKind      UpdateKind `json:"updateKind,omitempty"`
-	Error           string     `json:"error,omitempty"`
-}
-
-// RunStatus describes the most recent apply run, reconstructed from disk so
-// it stays accurate across the restart the update itself triggers.
-type RunStatus struct {
-	State        string     `json:"state"` // running | succeeded | failed
-	Target       string     `json:"target"`
-	UpdateKind   UpdateKind `json:"updateKind,omitempty"`
-	StartedAt    int64      `json:"startedAt"`
-	StartedBy    string     `json:"startedBy,omitempty"`
-	FinishedAt   int64      `json:"finishedAt,omitempty"`
-	ExitCode     *int       `json:"exitCode,omitempty"`
-	Log          string     `json:"log,omitempty"`
-	LogUpdatedAt int64      `json:"logUpdatedAt,omitempty"`
-	Progress     *Progress  `json:"progress,omitempty"`
-}
-
-// Progress is the durable, structured view of the updater's current phase.
-// The raw log remains available for diagnostics, while these fields let the
-// UI explain long-running work without guessing from terminal output.
-type Progress struct {
-	Phase       string `json:"phase"`
-	Message     string `json:"message"`
-	Completed   int    `json:"completed,omitempty"`
-	Total       int    `json:"total,omitempty"`
-	CurrentItem string `json:"currentItem,omitempty"`
-	UpdatedAt   int64  `json:"updatedAt"`
-}
-
-type Status struct {
-	CurrentVersion string       `json:"currentVersion"`
-	LastCheck      *CheckResult `json:"lastCheck,omitempty"`
-	Run            *RunStatus   `json:"run,omitempty"`
-}
-
-type runRecord struct {
-	Target     string     `json:"target"`
-	UpdateKind UpdateKind `json:"updateKind,omitempty"`
-	StartedAt  int64      `json:"startedAt"`
-	StartedBy  string     `json:"startedBy"`
-	PID        int        `json:"pid"`
-}
-
-type doneRecord struct {
-	ExitCode   int   `json:"exitCode"`
-	FinishedAt int64 `json:"finishedAt"`
 }
 
 // Status reports the running version, the last check result, and the most
@@ -203,14 +131,14 @@ func (s *Service) Apply(ctx context.Context, startedBy, tag string) (Status, err
 	}); err != nil {
 		return s.statusLocked(), err
 	}
-	pid, err := s.host.StartUpdater(
-		s.installDir,
-		tag,
-		string(kind),
-		s.logPath(),
-		s.donePath(),
-		s.progressPath(),
-	)
+	pid, err := s.host.StartUpdater(UpdaterLaunch{
+		InstallDir:   s.installDir,
+		Target:       tag,
+		Kind:         kind,
+		LogPath:      s.logPath(),
+		DonePath:     s.donePath(),
+		ProgressPath: s.progressPath(),
+	})
 	if err != nil {
 		_ = os.Remove(s.progressPath())
 		return s.statusLocked(), fmt.Errorf("start updater: %w", err)
